feat(sln): remove obsolete api references
