+++ Directory.Packages.props
@@ -6,12 +6,12 @@
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
   </PropertyGroup>
   <ItemGroup>
-    <PackageVersion Include="Kentico.Xperience.Admin" Version="29.5.0" />
-    <PackageVersion Include="Kentico.Xperience.Azurestorage" Version="29.5.0" />
-    <PackageVersion Include="Kentico.Xperience.Cloud" Version="29.5.0" />
-    <PackageVersion Include="Kentico.Xperience.Imageprocessing" Version="29.5.0" />
-    <PackageVersion Include="Kentico.Xperience.WebApp" Version="29.5.0" />
-    <PackageVersion Include="Kentico.Xperience.Core.Tests" Version="29.5.0" />
+    <PackageVersion Include="Kentico.Xperience.Admin" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.Azurestorage" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.Cloud" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.Imageprocessing" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.WebApp" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.Core.Tests" Version="29.5.3" />
     <PackageVersion Include="Algolia.Search" Version="6.16.0" />
     <PackageVersion Include="Microsoft.Extensions.Configuration.UserSecrets" Version="6.0.1" />
     <PackageVersion Include="SonarAnalyzer.CSharp" Version="9.6.0.74858" />
