+++ examples/DancingGoat/.config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "kentico.xperience.dbmanager": {
-      "version": "29.5.0",
+      "version": "29.5.3",
       "commands": [
         "kentico-xperience-dbmanager"
       ]
