+++ examples/DancingGoat/Components/ViewComponents/CafeCardSection/CafeCardSectionViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+using DancingGoat.Models;
+
+namespace DancingGoat.ViewComponents
+{
+    public record CafeCardSectionViewModel(IEnumerable<CafeViewModel> Cafes, string ContactsPagePath)
+    {
+    }
+}
