+++ examples/DancingGoat/Components/ViewComponents/TrackingConsent/TrackingConsentViewComponent.cs
@@ -54,7 +54,7 @@ namespace DancingGoat.ViewComponents
                 {
                     ConsentShortText = (await consent.GetConsentTextAsync(currentLanguage)).ShortText,
                     ReturnPageUrl = webPageDataContextRetriever.TryRetrieve(out var currentWebPageContext)
-                        ? (await urlRetriever.Retrieve(currentWebPageContext.WebPage.WebPageItemID, currentLanguage)).RelativePath
+                        ? (await urlRetriever.Retrieve(currentWebPageContext.WebPage.WebPageItemID, currentLanguage, cancellationToken: HttpContext.RequestAborted)).RelativePath
                         : (HttpContext.Request.PathBase + HttpContext.Request.Path).Value
                 };
 
@@ -62,7 +62,7 @@ namespace DancingGoat.ViewComponents
                 if ((contact != null) && consentAgreementService.IsAgreed(contact, consent))
                 {
                     consentModel.IsConsentAgreed = true;
-                    consentModel.PrivacyPageUrl = Url.Content((await urlRetriever.Retrieve(PrivacyPageConstants.PRIVACY_PAGE_TREE_PATH, websiteChannelContext.WebsiteChannelName, currentLanguage)).RelativePath);
+                    consentModel.PrivacyPageUrl = Url.Content((await urlRetriever.Retrieve(PrivacyPageConstants.PRIVACY_PAGE_TREE_PATH, websiteChannelContext.WebsiteChannelName, currentLanguage, cancellationToken: HttpContext.RequestAborted)).RelativePath);
                 }
 
                 return View("~/Components/ViewComponents/TrackingConsent/Default.cshtml", consentModel);
