+++ examples/DancingGoat/Models/Reusable/Cafe/CafeRepository.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 
 using CMS.ContentEngine;
 using CMS.Helpers;
-using CMS.Websites;
 using CMS.Websites.Routing;
 
 namespace DancingGoat.Models
@@ -30,13 +29,13 @@ namespace DancingGoat.Models
         }
 
         /// <summary>
-        /// Returns an enumerable collection of company cafes ordered by a position in the content tree.
+        /// Returns an enumerable collection of cafes.
         /// </summary>
-        public async Task<IEnumerable<Cafe>> GetCompanyCafes(int count, string languageName, CancellationToken cancellationToken = default)
+        public async Task<IEnumerable<Cafe>> GetCafes(int count, string languageName, CancellationToken cancellationToken = default)
         {
             var queryBuilder = GetQueryBuilder(count, languageName);
 
-            var cacheSettings = new CacheSettings(5, WebsiteChannelContext.WebsiteChannelName, nameof(CafeRepository), languageName, nameof(GetCompanyCafes), count);
+            var cacheSettings = new CacheSettings(5, WebsiteChannelContext.WebsiteChannelName, nameof(CafeRepository), nameof(GetCafes), count, languageName);
 
             return await GetCachedQueryResult<Cafe>(queryBuilder, null, cacheSettings, GetDependencyCacheKeys, cancellationToken);
         }
@@ -47,9 +46,8 @@ namespace DancingGoat.Models
             return new ContentItemQueryBuilder()
                     .ForContentType(Cafe.CONTENT_TYPE_NAME,
                         config => config
-                            .WithLinkedItems(1)
-                            .TopN(count)
-                            .Where(where => where.WhereTrue(nameof(Cafe.CafeIsCompanyCafe))))
+                        .WithLinkedItems(1)
+                        .TopN(count))
                     .InLanguage(languageName);
         }
