+++ examples/DancingGoat/Models/Reusable/Tag/TagViewModel.cs
@@ -6,39 +6,35 @@ using CMS.ContentEngine;
 
 using Tag = CMS.ContentEngine.Tag;
 
-namespace DancingGoat.Models
-{
-    public record TagViewModel(string Name, int Level, Guid Value, bool IsChecked = false)
-    {
-        private const int ROOT_TAG_ID = 0;
+namespace DancingGoat.Models;
 
-        public static TagViewModel GetViewModel(Tag tag, int level = 0)
+public record TagViewModel(string Name, int Level, Guid Value, bool IsChecked = false)
+{
+    private const int ROOT_TAG_ID = 0;
+
+    public static TagViewModel GetViewModel(Tag tag, int level = 0) => new TagViewModel(tag.Title, level, tag.Identifier);
+
+
+    public static List<TagViewModel> GetViewModels(IEnumerable<Tag> tags)
+    {
+        var result = new List<TagViewModel>();
+        var tagsByParentId = tags.GroupBy(tag => tag.ParentID).ToDictionary(group => group.Key, group => group.ToList());
+
+        if (tagsByParentId.TryGetValue(ROOT_TAG_ID, out var firstLevelTags))
         {
-            return new TagViewModel(tag.Title, level, tag.Identifier);
+            GetTagsWithTagViewModels(firstLevelTags, ROOT_TAG_ID);
         }
 
+        return result;
 
-        public static List<TagViewModel> GetViewModels(IEnumerable<Tag> tags)
+
+        void GetTagsWithTagViewModels(IEnumerable<Tag> currentLevelTags, int level)
         {
-            var result = new List<TagViewModel>();
-            var tagsByParentId = tags.GroupBy(tag => tag.ParentID).ToDictionary(group => group.Key, group => group.ToList());
-
-            if (tagsByParentId.TryGetValue(ROOT_TAG_ID, out var firstLevelTags))
+            foreach (var tag in currentLevelTags.OrderBy(tag => tag.Order))
             {
-                GetTagsWithTagViewModels(firstLevelTags, ROOT_TAG_ID);
-            }
-
-            return result;
-
-
-            void GetTagsWithTagViewModels(IEnumerable<Tag> currentLevelTags, int level)
-            {
-                foreach (var tag in currentLevelTags.OrderBy(tag => tag.Order))
-                {
-                    var children = tagsByParentId.TryGetValue(tag.ID, out var childrenTags) ? childrenTags : Enumerable.Empty<Tag>();
-                    result.Add(GetViewModel(tag, level));
-                    GetTagsWithTagViewModels(children, level + 1);
-                }
+                var children = tagsByParentId.TryGetValue(tag.ID, out var childrenTags) ? childrenTags : Enumerable.Empty<Tag>();
+                result.Add(GetViewModel(tag, level));
+                GetTagsWithTagViewModels(children, level + 1);
             }
         }
     }
