+++ examples/DancingGoat/Models/WebPage/ArticlePage/ArticleViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 using CMS.Websites;
@@ -11,11 +12,11 @@ namespace DancingGoat.Models
         /// <summary>
         /// Validates and maps <see cref="ArticlePage"/> to a <see cref="ArticleViewModel"/>.
         /// </summary>
-        public static async Task<ArticleViewModel> GetViewModel(ArticlePage articlePage, IWebPageUrlRetriever urlRetriever, string languageName)
+        public static async Task<ArticleViewModel> GetViewModel(ArticlePage articlePage, IWebPageUrlRetriever urlRetriever, string languageName, CancellationToken cancellationToken = default)
         {
             var teaser = articlePage.ArticlePageTeaser.FirstOrDefault();
 
-            var url = await urlRetriever.Retrieve(articlePage, languageName);
+            var url = await urlRetriever.Retrieve(articlePage, languageName, cancellationToken);
 
             return new ArticleViewModel(
                 articlePage.ArticleTitle,
