+++ examples/DancingGoat/Models/WebPage/ContactsPage/ContactsIndexViewModel.cs
@@ -18,6 +18,12 @@ namespace DancingGoat.Models
         public List<CafeViewModel> CompanyCafes { get; set; }
 
 
+        /// <summary>
+        /// The partner cafes data.
+        /// </summary>
+        public List<CafeViewModel> PartnerCafes { get; set; }
+
+
         /// <inheritdoc/>
         public IWebPageFieldsSource WebPage { get; init; }
     }
