+++ examples/DancingGoat/appsettings.json
@@ -19,8 +19,8 @@
   "CMSHashStringSalt": "s",
   "WebCrawlerBaseUrl": "http://localhost:15037",
   "CMSAlgoliaSearch": {
-    "ApplicationId": "Q2YK718J0J",
-    "ApiKey": "e394c4ceb17da01378f6924ab603c5f1",
-    "SearchKey": "7f14c628ce271806397ca4cae3b71e16"
+    "ApplicationId": "",
+    "ApiKey": "",
+    "SearchKey": ""
   }
 }
