+++ examples/DancingGoat/wwwroot/Content/Styles/Site.css
@@ -6428,6 +6428,7 @@ a + .article-tile-info .article-tile-date {
 /*# Taste our coffee #*/
 .ourcoffee-tile-image {
   overflow: hidden;
+  height: 100%;
   width: 100%;
 }
 .ourcoffee-tile-text {
