+++ examples/DancingGoat/wwwroot/Content/Styles/Site.less
@@ -1236,6 +1236,7 @@ a + .article-tile-info .article-tile-date {
 /*# Taste our coffee #*/
 .ourcoffee-tile-image {
     overflow: hidden;
+    height: 100%;
     width: 100%;
 }
