+++ src/Kentico.Xperience.Algolia/Admin/Client/package.json
@@ -13,8 +13,8 @@
     "node": ">=18.12.0 <=19"
   },
   "dependencies": {
-    "@kentico/xperience-admin-base": "29.5.0",
-    "@kentico/xperience-admin-components": "29.5.0",
+    "@kentico/xperience-admin-base": "29.5.3",
+    "@kentico/xperience-admin-components": "29.5.3",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-icons": "5.2.1",
@@ -43,7 +43,7 @@
     "eslint-plugin-react-hooks": "4.6.0",
     "prettier": "3.2.4",
     "typescript": "5.3.3",
-    "webpack": "5.90.0",
+    "webpack": "^5.95.0",
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "4.15.1"
   },
